pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localforge"
version = "0.1.0"
description = "A local-first, repo-aware coding agent powered by Ollama"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "localforge contributors" },
]
keywords = ["ai", "coding-agent", "ollama", "local-first"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer[all]",
    "rich",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx",
    "pyyaml",
    "pathspec",
    "tiktoken",
    "gitpython",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
localforge = "localforge.cli.main:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
